+++ convert/src/main/kotlin/com/github/zeldigas/text2confl/convert/markdown/MarkdownAttachmentCollector.kt
@@ -8,7 +8,7 @@ import com.vladsch.flexmark.util.ast.Document
 import com.vladsch.flexmark.util.ast.Node
 import com.vladsch.flexmark.util.ast.NodeVisitor
 import com.vladsch.flexmark.util.ast.VisitHandler
-import mu.KotlinLogging
+import io.github.oshai.kotlinlogging.KotlinLogging
 import java.nio.file.Path
 
 class MarkdownAttachmentCollector(