+++ core/src/main/kotlin/com/github/zeldigas/text2confl/core/upload/ContentUploader.kt
@@ -6,11 +6,11 @@ import com.github.zeldigas.text2confl.convert.EditorVersion
 import com.github.zeldigas.text2confl.convert.Page
 import com.github.zeldigas.text2confl.convert.PageHeader
 import com.github.zeldigas.text2confl.core.config.Cleanup
+import io.github.oshai.kotlinlogging.KotlinLogging
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.launch
-import mu.KotlinLogging
 
 
 class ContentUploader(